+++ java-client/src/main/java/org/opensearch/client/opensearch/_types/aggregations/MultiBucketAggregateBase.java
@@ -75,7 +75,7 @@ public abstract class MultiBucketAggregateBase<TBucket> extends AggregateBase {
 
     }
 
-    protected abstract static class AbstractBuilder<TBucket, BuilderT extends AbstractBuilder<TBucket, BuilderT>> extends
+    public abstract static class AbstractBuilder<TBucket, BuilderT extends AbstractBuilder<TBucket, BuilderT>> extends
         AggregateBase.AbstractBuilder<BuilderT> {
         private Buckets<TBucket> buckets;