+++ java-client/src/main/java/org/opensearch/client/opensearch/_types/aggregations/StatsAggregate.java
@@ -243,7 +243,7 @@ public class StatsAggregate extends AggregateBase implements AggregateVariant {
         }
     }
 
-    protected abstract static class AbstractBuilder<BuilderT extends AbstractBuilder<BuilderT>> extends AggregateBase.AbstractBuilder<
+    public abstract static class AbstractBuilder<BuilderT extends AbstractBuilder<BuilderT>> extends AggregateBase.AbstractBuilder<
         BuilderT> {
         private Long count;